haha wow, finish part 1
got terribly sidetracked on output that wasn't deterministic; tried adding locking and using the race detector, nothing. it's only one thread. walked away and remembered that ranging over a map is defined to have a nondeterministic walk order, and I was indirectly breaking out of a range based on a key value. whoops.