part 2 - correct but not sufficient
So this solution works for small sets, but is too slow for the actual input set. Another variation I tried was parallel, but I was not able to avoid memory exhaustion.

My next approach will be to try modeling the list as a network of nodes, connecting to every other "reachable" node, and see if I can collapse some of the work down to multiplication rather than just counting.